e2e logging and reconfiguration working
