+++ tests/reconfiguration.py
@@ -3,7 +3,6 @@
 import sys
 import e2e_args
 import infra.ccf
-import infra.proc
 import json
 
 import logging
@@ -14,7 +13,7 @@ from loguru import logger as LOG
 
 def check_can_progress(node):
     with node.node_client() as mc:
-        check_commit = infra.ccf.Checker(mc)
+        check_commit = infra.checker.Checker(mc)
         with node.node_client() as c:
             check_commit(c.rpc("mkSign", params={}), result=True)
 
@@ -69,8 +68,9 @@ def test_add_node_untrusted_code(network, args):
 # Network may need recovery after this test
 def test_retire_node(network, args):
     LOG.info("Retiring a backup")
+    primary, _ = network.find_primary()
     backup_to_retire = network.find_any_backup()
-    network.retire_node(backup_to_retire)
+    network.consortium.retire_node(primary, backup_to_retire)
     backup_to_retire.stop()
     return network
